refactor(s3): Change offset types to uint64 and add configurable retry
- Change WriteAt and ReadAt offset parameter from int64 to uint64
  across all content store interfaces and implementations
- Remove unnecessary negative offset validation (now enforced by type)
- Add retry with exponential backoff for transient S3 errors:
  - Network errors, throttling (429), server errors (5xx)
  - Configurable via YAML: max_retries, initial_backoff, max_backoff,
    backoff_multiplier
  - Non-retryable errors (404, 403, invalid range) fail immediately
- Update NFS handlers to use uint64 offsets directly without casting
- Add retry configuration documentation to CONFIGURATION.md

+++ internal/protocol/nfs/v3/handlers/commit.go
@@ -494,8 +494,7 @@ func flushCacheToContentStore(
 		return fmt.Errorf("cache read error: %w", err)
 	}
 
-	// Content store WriteAt uses int64 for io.ReaderAt compatibility
-	err = contentStore.WriteAt(ctx.Context, contentID, buf[:n], int64(flushedOffset))
+	err = contentStore.WriteAt(ctx.Context, contentID, buf[:n], flushedOffset)
 	if err != nil {
 		return fmt.Errorf("content store write error: %w", err)
 	}
